backend/requirements.txt
django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
razorpay
django-storages[azure]
psycopg2-binary
pillow
python-dotenv
dj-database-url
